Conflicts:
	plugins/Polls/class.polls.plugin.php

+++ plugins/Reactions/class.reactionmodel.php
@@ -119,7 +119,7 @@ class ReactionModel {
       $Data['TagID'] = $TagID;
       
       $Row = array();
-      $Columns = array('UrlCode', 'Name', 'Description', 'Sort', 'Class', 'TagID', 'Active');
+      $Columns = array('UrlCode', 'Name', 'Description', 'Sort', 'Class', 'TagID', 'Active', 'Custom');
       foreach ($Columns as $Column) {
          if (isset($Data[$Column])) {
             $Row[$Column] = $Data[$Column];
@@ -127,8 +127,26 @@ class ReactionModel {
          }
       }
       
+      // Check to see if the reaction type has been customized.
+      if (!isset($Row['Custom'])) {
+         $Current = self::ReactionTypes($UrlCode);
+         
+         if ($Current && GetValue('Custom', $Current))
+            return;
+      }
+      
+      // Check to see if the reaction type has been customized.
+      if (!isset($Row['Custom'])) {
+         $CurrentCustom = $this->SQL->GetWhere('ReactionType', array('UrlCode' => $UrlCode))->Value('Custom');
+         
+         if ($CurrentCustom)
+            return;
+      }
+      
       if (!empty($Data)) {
          $Row['Attributes'] = serialize($Data);
+      } else {
+         $Row['Attributes'] = NULL;
       }
       
       Gdn::SQL()->Replace('ReactionType', $Row, array('UrlCode' => $UrlCode), TRUE);
@@ -231,7 +249,7 @@ class ReactionModel {
       return array($Row, $Model, $Log);
    }
    
-   public function JoinUserTags(&$Data, $RecordType) {
+   public function JoinUserTags(&$Data, $RecordType = FALSE) {
       if (!$Data)
          return;
       
@@ -248,36 +266,56 @@ class ReactionModel {
 //      decho($Data);
       
       foreach ($Data2 as $Row) {
-         $ID = GetValue($PK, $Row);
+         if (!$RecordType)
+            $RT = GetValue('RecordType', $Row);
+         else
+            $RT = $RecordType;
+         
+         $ID = GetValue($RT.'ID', $Row);
+         
          if ($ID)
-            $IDs[$ID] = 1;
+            $IDs[$RT][$ID] = 1;
       }
 //      decho($IDs);
       
-      $TagsData = $this->SQL
-         ->Select('RecordID')
-         ->Select('UserID')
-         ->Select('TagID')
-         ->Select('DateInserted')
-         ->From('UserTag')
-         ->Where('RecordType', $RecordType)
-         ->WhereIn('RecordID', array_keys($IDs))
-         ->OrderBy('DateInserted')
-         ->Get()->ResultArray();
-      
-      
+      $TagsData = array();
+      foreach ($IDs as $RT => $In) {
+         $TagsData[$RT] = $this->SQL
+            ->Select('RecordID')
+            ->Select('UserID')
+            ->Select('TagID')
+            ->Select('DateInserted')
+            ->From('UserTag')
+            ->Where('RecordType', $RT)
+            ->WhereIn('RecordID', array_keys($In))
+            ->OrderBy('DateInserted')
+            ->Get()->ResultArray();
+      }
       
       $Tags = array();
-      foreach($TagsData as $Row) {
-         $UserIDs[$Row['UserID']] = 1;
-         $Tags[$Row['RecordID']][] = $Row;
+      foreach($TagsData as $RT => $Rows) {
+         foreach ($Rows as $Row) {
+            $UserIDs[$Row['UserID']] = 1;
+            $Tags[$RT.'-'.$Row['RecordID']][] = $Row;
+         }
       }
+      
+//      decho($Tags, 'Tags');
+//      die();
       
       // Join the tags.
       foreach ($Data2 as &$Row) {
+         if ($RecordType)
+            $RT = $RecordType;
+         else
+            $RT = GetValue('RecordType', $Row);
+         if (!$RT)
+            $RT = 'RecordType';
+         $PK = $RT.'ID';
          $ID = GetValue($PK, $Row);
+         
          if ($ID)
-            $TagRow = GetValue($ID, $Tags, array());
+            $TagRow = GetValue($RT.'-'.$ID, $Tags, array());
          else
             $TagRow = array();
          
@@ -589,9 +627,14 @@ class ReactionModel {
       $LogThreshold = GetValue('LogThreshold', $ReactionType, 10000000);
       $RemoveThreshold = GetValue('RemoveThreshold', $ReactionType, 10000000);
       
-      if (!GetValueR($AttrColumn.'.RestoreUserID', $Row)) {
+      if (!GetValueR($AttrColumn.'.RestoreUserID', $Row) || Debug()) {
          // We are only going to remove stuff if the record has not been verified.
          $Log = GetValue('Log', $ReactionType, 'Moderation');
+         
+         // Do a sanity check to not delete too many comments.
+         $NoDelete = FALSE;
+         if ($RecordType == 'Discussion' && $Row['CountComments'] > 3)
+            $NoDelete = TRUE;
          
          $LogOptions = array('GroupBy' => array('RecordID'));
          $UndoButton = '';
@@ -601,14 +644,14 @@ class ReactionModel {
             $OtherUserData = $this->SQL->GetWhere('UserTag', array('RecordType' => $RecordType, 'RecordID' => $ID, 'TagID' => $ReactionType['TagID']))->ResultArray();
             $OtherUserIDs = array();
             foreach ($OtherUserData as $UserRow) {
-               if ($UserRow['UserID'] == $UserID || !$Row['UserID'])
+               if ($UserRow['UserID'] == $UserID || !$UserRow['UserID'])
                   continue;
                $OtherUserIDs[] = $UserRow['UserID'];
             }
             $LogOptions['OtherUserIDs'] = $OtherUserIDs;
          }
          
-         if ($Score >= $RemoveThreshold) {
+         if (!$NoDelete && $Score >= $RemoveThreshold) {
             // Remove the record to the log.
             $Model->Delete($ID, array('Log' => $Log, 'LogOptions' => $LogOptions));
             $Message = array(
